Initial upload
I have been listening to a lot of John Prine's songs since he passed away a week or so ago. It is revitalizing my interest in playing and singing, so that is something to be thankful for along with his seemingly endless body of work.

A long time ago I created a C++ program to convert the format of chords from above the lyrics to within the lyrics. This takes up less space and I can get through a song with less scrolling. I lost that program so I wrote this one over the last couple of days, this time in python. 
Example:
On a mac or linux system (not sure about windows):
python helloInThere.txt > hello.txt
hello.txt will contain the same song with chords in square brackets embedded in the line at approximately the same location as the original . Sometimes things get whacky because tabs are converted to various widths and so forth. But this works pretty well at identify the chord lines and inserting the chords. 

I have been saving the re-formatted songs in a google doc: 
https://docs.google.com/document/d/1OFT5wi4eDwbqVMn6keq9a5aZOaHFIKL1AZ6wxDMWs3c/edit?usp=sharing