Add files via upload
Added convert.sh that converts all the *txt file in folder "orig" and  puts the chord-in-lyric output in the "converted" folder
Also made some changes to chordpro.py. As I convert more files, problems sometimes arise. In this case it was how to handle
consecutive chord lines that are not followed by a lyric line. The new version will just print the chords out in the right order.